feat(authentication): add authentication functionality
Currently, the program has no functionality for logging in or registering
users.

This commit adds this ability, as it gives users an option to log in and
register. It adds a modal for both the functionality and displays a users
name on the navbar when they have logged in succesfully.

It adds services that can be used by a to check if there exists a user,
and to get the currently logged in user. It also provides a service that
can be used to open modals.

[Finishes #153976523]