chore(modal): redesign modals and auth button
This commit redesigns the modals in the applications to make them larger.
It also transfers the login button to the top right corner of the
screen which is to be replaced by the user name once user has logged in
succesfully.

It also adds the option to register on the login modal, and to login
on the register modal.

[Delivers 155120497]