#155552251 Handle error states (#7)
* WIP

* ft(error): add way to handle errors states

This commit adds a 404, and 500 pages to be displayed if user
attempts to access wrong page.

It adds a pop up displaying error encountered if any.

[Finishes #155552251]

* ft(error-handler): adds handlers for errors.

This commit adds the functionality to log user out automatically
if the server responds with a 404. It also adds the ability to route
to a 404 page if the user attempts to route to a page that doesn't
exist.

[Finishes #155552251]

* feat(error-handler): adds background images to error pages

Adds different background images to 404 and 500 error pages.

[Finishes #155552251]

* feat(items-table): redesign items table

This commit redesigns the lost and found page to give more room
for the items. It moves the add item button to the top of the
page together with the section switchers.

It also make the button disappear if the page is minimised to less
than 700 px.

[Finishes #155552251]