feat(add-button): add an add item button
This commit gives the add item button ability to change color and
activeness whether a user is logged in or logged out. It turns to
gray and is disabled if the user is not logged out, and to green if
user is logged in. It also adds custom tool tip text to the button
depending on whether the user is logged in.

It also adds ability to theme the items to give them a compact or
classic layout, compact making the items shorter an thinner thus
fitting more items in to little space and classic wider and longer,
thus fits fewer items in single space.

[Finishes #155949493]