#155949493 Add 'add item' button (#9)
* feat(add-item): add the button to add item

This commit adds the button to add items as a circular button
at the bottom right of the screen. It also changes some styles for
the modal.

[Finishes #155949493]

* feat(add-button): adds close button for modals

This commit adds a close button for modals which is on the left side
when window width is more than 700px but goes to the right side
when the window width is less than 700px.

[Finishes #155949493]

* feat(add-button): add an add item button

This commit gives the add item button ability to change color and
activeness whether a user is logged in or logged out. It turns to
gray and is disabled if the user is not logged out, and to green if
user is logged in. It also adds custom tool tip text to the button
depending on whether the user is logged in.

It also adds ability to theme the items to give them a compact or
classic layout, compact making the items shorter an thinner thus
fitting more items in to little space and classic wider and longer,
thus fits fewer items in single space.

[Finishes #155949493]