#156872933 Add animations to modal and lost and found (#10)
* feat(add-item): add the button to add item

This commit adds the button to add items as a circular button
at the bottom right of the screen. It also changes some styles for
the modal.

[Finishes #155949493]

* feat(add-button): add an add item button

This commit gives the add item button ability to change color and
activeness whether a user is logged in or logged out. It turns to
gray and is disabled if the user is not logged out, and to green if
user is logged in. It also adds custom tool tip text to the button
depending on whether the user is logged in.

It also adds ability to theme the items to give them a compact or
classic layout, compact making the items shorter an thinner thus
fitting more items in to little space and classic wider and longer,
thus fits fewer items in single space.

[Finishes #155949493]

* chore(intergration-tests): fix intergration tests

This commit fixes intergration tests and ensures all intergration tests
for components are passing.

[Delivers #156025629]

* feat(test): adds tests for items module

This commit adds tests for the items module components. It attempts
to fully test all the component functionalities to ensure behaviour
is as expected.

It also adds tests files to all services and ensure they are passing
to set a stepping stone for further testing of the services.

[Finishes #156025629]

* feat(animations): adds animations to page

This commit adds animations on lost and found component. It also
adds fade in animation to router outlet.

* feat(animations): adds angular animations to pages.

This commit adds animations to the modal component, and the lost
and found component.

It also adds the BrowserAnimationsModule to prevent failing tests.