#156025629 Add integration tests (#11)
* chore(intergration-tests): fix intergration tests

This commit fixes intergration tests and ensures all intergration tests
for components are passing.

[Delivers #156025629]

* feat(test): adds tests for items module

This commit adds tests for the items module components. It attempts
to fully test all the component functionalities to ensure behaviour
is as expected.

It also adds tests files to all services and ensure they are passing
to set a stepping stone for further testing of the services.

[Finishes #156025629]

* feat(tests): add tests for shared module components

* chore(round-button): makes round button component reusable

This commit makes changes the name of addbuttoncomponent to round
button component to make the component independent thus easily
reusable. It give responsibility to tilt the button to the
button itself with an Input binding field and make it a close
button.